+++ src/snowflake/cli/api/identifiers.py
@@ -27,6 +27,7 @@ from snowflake.cli.api.project.schemas.v1.identifier_model import (
 from snowflake.cli.api.project.util import (
     VALID_IDENTIFIER_REGEX,
     identifier_for_url,
+    sanitize_identifier,
     unquote_identifier,
 )
 
@@ -179,10 +180,15 @@ class FQN:
     ) -> "FQN":
         """Create an instance related to another Snowflake resource."""
         unquoted_name = unquote_identifier(resource_fqn.name)
+        safe_name = sanitize_identifier(unquoted_name)
         safe_cli_name = resource_type.value.cli_name.upper().replace("-", "_")
-        return cls.from_string(
-            f"{safe_cli_name}_{unquoted_name}_{int(time.time())}_{purpose.upper()}"
-        ).using_context()
+        fqn = cls.from_string(
+            f"{safe_cli_name}_{safe_name}_{int(time.time())}_{purpose.upper()}"
+        )
+        fqn.set_database(resource_fqn.database)
+        fqn.set_schema(resource_fqn.schema)
+        fqn.using_context()
+        return fqn
 
     def set_database(self, database: str | None) -> "FQN":
         if database: